+++ examples/zephyr/matrix_multiply_rpmsglite/remote/src/main.cpp
@@ -15,7 +15,7 @@
 /*******************************************************************************
  * Definitions
  ******************************************************************************/
-#define ERPC_TRANSPORT_RPMSG_LITE_LINK_ID (RL_PLATFORM_IMXRT1170_M7_M4_LINK_ID)
+#define ERPC_TRANSPORT_RPMSG_LITE_LINK_ID (CONFIG_RPMSG_LITE_LINK_ID)
 
 #define SHM_MEM_ADDR DT_REG_ADDR(DT_CHOSEN(zephyr_ipc_shm))
 #define SHM_MEM_SIZE DT_REG_SIZE(DT_CHOSEN(zephyr_ipc_shm))