+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Zephyr examples added support for RT600 and RT700.
+
 ### Updated
 
 - eRPC Zephyr module port updated for Zephyr version 4.4